Fix bug in vllmrerank and sentransformerrerank
