+++ xinference/model/rerank/tests/test_rerank.py
@@ -20,58 +20,6 @@ import tempfile
 import pytest
 
 from ....client import Client
-from ..cache_manager import RerankCacheManager
-from ..core import RerankModelFamilyV2, RerankSpecV1
-from ..sentence_transformers.core import SentenceTransformerRerankModel
-
-TEST_MODEL_SPEC = RerankModelFamilyV2(
-    version=2,
-    model_name="bge-reranker-base",
-    type="normal",
-    max_tokens=512,
-    language=["en", "zh"],
-    model_specs=[
-        RerankSpecV1(
-            model_id="bge-reranker-base",
-            model_revision="465b4b7ddf2be0a020c8ad6e525b9bb1dbb708ae",
-            model_format="pytorch",
-        )
-    ],
-)
-
-
-def test_model():
-    model_path = None
-    try:
-        model_path = RerankCacheManager(TEST_MODEL_SPEC).cache()
-        model = SentenceTransformerRerankModel("mock", model_path, TEST_MODEL_SPEC)
-
-        query = "A man is eating pasta."
-        # With all sentences in the corpus
-        corpus = [
-            "A man is eating food.",
-            "A man is eating a piece of bread.",
-            "The girl is carrying a baby.",
-            "A man is riding a horse.",
-            "A woman is playing violin.",
-            "Two men pushed carts through the woods.",
-            "A man is riding a white horse on an enclosed ground.",
-            "A monkey is playing drums.",
-            "A cheetah is running behind its prey.",
-        ]
-        model.load()
-        scores = model.rerank(corpus, query, None, None, True, True)
-        assert scores["results"][0]["index"] == 0
-        assert scores["results"][0]["document"]["text"] == corpus[0]
-
-        n_tokens = scores["meta"]["tokens"]["input_tokens"]
-        tokenizer = model._model.tokenizer
-        expect_n_tokens = sum(len(tokenizer.tokenize([query, d])) for d in corpus)
-        assert n_tokens >= expect_n_tokens
-
-    finally:
-        if model_path is not None:
-            shutil.rmtree(model_path, ignore_errors=True)
 
 
 @pytest.mark.parametrize("model_name", ["bge-reranker-base", "bge-reranker-v2-m3"])
